Solucionados errores que no permitían al algoritmo de Douglas Peucker calcular distancias de forma correcta. Solucionados errores en el script de compilación que impedían que se entrenara la librería con los nuevos ficheros generados por la aplicación de simplificación.

+++ test/simplificacionSTASM/inc/douglasPeucker.h
@@ -4,7 +4,7 @@
 #include <opencv2/opencv.hpp>
 
 // Función que calcula la distancia de un punto _pt a la linea formada por 'l1' y 'l2'.
-double distanciaPerpendicular(cv::Point2i _pt, cv::Point2i l1, cv::Point2i l2);
+double distanciaPerpendicular(cv::Point2f _pt, cv::Point2f l1, cv::Point2f l2);
 
 // Función que implementa el algoritmo de Douglas Peucker.
 // https://en.wikipedia.org/wiki/Ramer%E2%80%93Douglas%E2%80%93Peucker_algorithm
